move ASSIST to constant
