+++ openpype/hosts/nuke/api/constants.py
@@ -1,0 +1,4 @@
+import os
+
+
+ASSIST = bool(os.getenv("NUKEASSIST"))
